backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
pydantic==2.7.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
